# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GDGOC Hacktoberfest Backend",
    "description": "Event-management API for GDGOC Hacktoberfest open-source contribution tracking.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/dashboard/student": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Student dashboard overview",
        "description": "Personalized student dashboard: active claimed issues, PR progress, and valid contribution count.",
        "operationId": "student_dashboard_dashboard_student_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/maintainer": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Maintainer review queue dashboard",
        "description": "Maintainer dashboard: review queue of open PRs, pending reviews, and recent review activity.",
        "operationId": "maintainer_dashboard_dashboard_maintainer_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintainerDashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/repository/{repository_id}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Repository statistics dashboard",
        "description": "Repository-level dashboard: issue breakdown, PR counts, commit velocity, and contributor count.",
        "operationId": "repository_dashboard_dashboard_repository__repository_id__get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryDashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActiveClaimSummary": {
        "properties": {
          "claim_id": {
            "type": "integer",
            "title": "Claim Id"
          },
          "issue_id": {
            "type": "integer",
            "title": "Issue Id"
          },
          "issue_title": {
            "type": "string",
            "title": "Issue Title"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "repo_name": {
            "type": "string",
            "title": "Repo Name"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "claimed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Claimed At"
          }
        },
        "type": "object",
        "required": [
          "claim_id",
          "issue_id",
          "issue_title",
          "difficulty",
          "repo_name",
          "platform",
          "claimed_at"
        ],
        "title": "ActiveClaimSummary"
      },
...
      "ClaimUserBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "psit_roll_no": {
            "type": "string",
            "title": "Psit Roll No"
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "psit_roll_no"
        ],
        "title": "ClaimUserBrief"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkedIssueBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "github_issue_id": {
            "type": "integer",
            "title": "Github Issue Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "github_issue_id",
          "title",
          "difficulty",
          "status"
        ],
        "title": "LinkedIssueBrief"
      },
...
      "MaintainerDashboardResponse": {
        "properties": {
          "maintainer_id": {
            "type": "integer",
            "title": "Maintainer Id"
          },
          "pending_reviews_count": {
            "type": "integer",
            "title": "Pending Reviews Count"
          },
          "approved_prs_count": {
            "type": "integer",
            "title": "Approved Prs Count"
          },
          "changes_requested_count": {
            "type": "integer",
            "title": "Changes Requested Count"
          },
          "review_queue": {
            "items": {
              "$ref": "#/components/schemas/ReviewQueueItem"
            },
            "type": "array",
            "title": "Review Queue",
            "default": []
          },
          "recently_reviewed": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recently Reviewed",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "maintainer_id",
          "pending_reviews_count",
          "approved_prs_count",
          "changes_requested_count"
        ],
        "title": "MaintainerDashboardResponse"
      },
...
      "PRResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "repo_id": {
            "type": "integer",
            "title": "Repo Id"
          },
          "github_pr_id": {
            "type": "integer",
            "title": "Github Pr Id"
          },
          "issue_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/PRStatus"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimUserBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimUserBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "repository": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepositoryBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "linked_issue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LinkedIssueBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewDetail"
            },
            "type": "array",
            "title": "Reviews",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "repo_id",
          "github_pr_id",
          "user_id",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "PRResponse"
      },
      "PRStatus": {
        "type": "string",
        "enum": [
          "open",
          "merged",
          "closed",
          "draft"
        ],
        "title": "PRStatus"
      },
      "RepositoryBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "github_repo_url": {
            "type": "string",
            "title": "Github Repo Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "platform",
          "github_repo_url"
        ],
        "title": "RepositoryBrief"
      },
      "RepositoryDashboardResponse": {
        "properties": {
          "repository": {
            "$ref": "#/components/schemas/RepositoryBrief"
          },
          "total_issues": {
            "type": "integer",
            "title": "Total Issues"
          },
          "open_issues": {
            "type": "integer",
            "title": "Open Issues"
          },
          "claimed_issues": {
            "type": "integer",
            "title": "Claimed Issues"
          },
          "closed_issues": {
            "type": "integer",
            "title": "Closed Issues"
          },
          "total_prs": {
            "type": "integer",
            "title": "Total Prs"
          },
          "open_prs": {
            "type": "integer",
            "title": "Open Prs"
          },
          "merged_prs": {
            "type": "integer",
            "title": "Merged Prs"
          },
          "total_commits": {
            "type": "integer",
            "title": "Total Commits"
          },
          "unique_contributors_count": {
            "type": "integer",
            "title": "Unique Contributors Count"
          },
          "recent_activity": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent Activity",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "repository",
          "total_issues",
          "open_issues",
          "claimed_issues",
          "closed_issues",
          "total_prs",
          "open_prs",
          "merged_prs",
          "total_commits",
          "unique_contributors_count"
        ],
        "title": "RepositoryDashboardResponse"
      },
      "ReviewDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pr_id": {
            "type": "integer",
            "title": "Pr Id"
          },
          "reviewer_id": {
            "type": "integer",
            "title": "Reviewer Id"
          },
          "reviewer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Name"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pr_id",
          "reviewer_id",
          "status",
          "reviewed_at"
        ],
        "title": "ReviewDetail"
      },
      "ReviewQueueItem": {
        "properties": {
          "pr_id": {
            "type": "integer",
            "title": "Pr Id"
          },
          "github_pr_id": {
            "type": "integer",
            "title": "Github Pr Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "repo_name": {
            "type": "string",
            "title": "Repo Name"
          },
          "contributor_name": {
            "type": "string",
            "title": "Contributor Name"
          },
          "contributor_roll_no": {
            "type": "string",
            "title": "Contributor Roll No"
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Status"
          }
        },
        "type": "object",
        "required": [
          "pr_id",
          "github_pr_id",
          "title",
          "repo_name",
          "contributor_name",
          "contributor_roll_no",
          "created_at",
          "status"
        ],
        "title": "ReviewQueueItem"
      },
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "approved",
          "changes_requested",
          "commented",
          "dismissed"
        ],
        "title": "ReviewStatus"
      },
...
      "StudentDashboardResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "psit_roll_no": {
            "type": "string",
            "title": "Psit Roll No"
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "active_claims_count": {
            "type": "integer",
            "title": "Active Claims Count"
          },
          "prs_submitted_count": {
            "type": "integer",
            "title": "Prs Submitted Count"
          },
          "prs_merged_count": {
            "type": "integer",
            "title": "Prs Merged Count"
          },
          "valid_contributions_count": {
            "type": "integer",
            "title": "Valid Contributions Count"
          },
          "active_claims": {
            "items": {
              "$ref": "#/components/schemas/ActiveClaimSummary"
            },
            "type": "array",
            "title": "Active Claims",
            "default": []
          },
          "recent_prs": {
            "items": {
              "$ref": "#/components/schemas/PRResponse"
            },
            "type": "array",
            "title": "Recent Prs",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "psit_roll_no",
          "verified",
          "active_claims_count",
          "prs_submitted_count",
          "prs_merged_count",
          "valid_contributions_count"
        ],
        "title": "StudentDashboardResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}